+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.litecoin_balance",
+    package = "com.coinhark.litecoinbalance",
 )
 
 android_resource(
     name = "res",
-    package = "com.litecoin_balance",
+    package = "com.coinhark.litecoinbalance",
     res = "src/main/res",
 )
