+++ android/app/build.gradle
@@ -98,7 +98,7 @@ android {
     buildToolsVersion '26.0.2'
 
     defaultConfig {
-        applicationId "com.litecoin_balance"
+        applicationId "com.coinhark.litecoinbalance"
         minSdkVersion 16
         targetSdkVersion 22
         versionCode 1
