+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">litecoin_balance</string>
+    <string name="app_name">Litecoin Balance</string>
 </resources>
