+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'litecoin_balance'
+rootProject.name = 'LitecoinBalance'
 include ':react-native-camera'
 project(':react-native-camera').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-camera/android')
 include ':react-native-vector-icons'
