+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "litecoin_balance",
-  "displayName": "litecoin_balance"
+  "name": "LitecoinBalance",
+  "displayName": "Litecoin Balance"
 }
