+++ index.js
@@ -1,4 +1,4 @@
 import { AppRegistry } from 'react-native';
 import App from './App';
 
-AppRegistry.registerComponent('litecoin_balance', () => App);
+AppRegistry.registerComponent('LitecoinBalance', () => App);
