+++ ios/LitecoinBalance-tvOS/Info.plist
