+++ ios/LitecoinBalance-tvOSTests/Info.plist
