+++ ios/LitecoinBalance.xcodeproj/xcshareddata/xcschemes/LitecoinBalance-tvOS.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "litecoin_balance-tvOS.app"
-               BlueprintName = "litecoin_balance-tvOS"
-               ReferencedContainer = "container:litecoin_balance.xcodeproj">
+               BuildableName = "LitecoinBalance-tvOS.app"
+               BlueprintName = "LitecoinBalance-tvOS"
+               ReferencedContainer = "container:LitecoinBalance.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "litecoin_balance-tvOSTests.xctest"
-               BlueprintName = "litecoin_balance-tvOSTests"
-               ReferencedContainer = "container:litecoin_balance.xcodeproj">
+               BuildableName = "LitecoinBalance-tvOSTests.xctest"
+               BlueprintName = "LitecoinBalance-tvOSTests"
+               ReferencedContainer = "container:LitecoinBalance.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "litecoin_balance-tvOSTests.xctest"
-               BlueprintName = "litecoin_balance-tvOSTests"
-               ReferencedContainer = "container:litecoin_balance.xcodeproj">
+               BuildableName = "LitecoinBalance-tvOSTests.xctest"
+               BlueprintName = "LitecoinBalance-tvOSTests"
+               ReferencedContainer = "container:LitecoinBalance.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "litecoin_balance-tvOS.app"
-            BlueprintName = "litecoin_balance-tvOS"
-            ReferencedContainer = "container:litecoin_balance.xcodeproj">
+            BuildableName = "LitecoinBalance-tvOS.app"
+            BlueprintName = "LitecoinBalance-tvOS"
+            ReferencedContainer = "container:LitecoinBalance.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "litecoin_balance-tvOS.app"
-            BlueprintName = "litecoin_balance-tvOS"
-            ReferencedContainer = "container:litecoin_balance.xcodeproj">
+            BuildableName = "LitecoinBalance-tvOS.app"
+            BlueprintName = "LitecoinBalance-tvOS"
+            ReferencedContainer = "container:LitecoinBalance.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "litecoin_balance-tvOS.app"
-            BlueprintName = "litecoin_balance-tvOS"
-            ReferencedContainer = "container:litecoin_balance.xcodeproj">
+            BuildableName = "LitecoinBalance-tvOS.app"
+            BlueprintName = "LitecoinBalance-tvOS"
+            ReferencedContainer = "container:LitecoinBalance.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
