+++ ios/LitecoinBalance.xcodeproj/xcshareddata/xcschemes/LitecoinBalance.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "litecoin_balance.app"
-               BlueprintName = "litecoin_balance"
-               ReferencedContainer = "container:litecoin_balance.xcodeproj">
+               BuildableName = "LitecoinBalance.app"
+               BlueprintName = "LitecoinBalance"
+               ReferencedContainer = "container:LitecoinBalance.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "litecoin_balanceTests.xctest"
-               BlueprintName = "litecoin_balanceTests"
-               ReferencedContainer = "container:litecoin_balance.xcodeproj">
+               BuildableName = "LitecoinBalanceTests.xctest"
+               BlueprintName = "LitecoinBalanceTests"
+               ReferencedContainer = "container:LitecoinBalance.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -62,9 +62,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "litecoin_balanceTests.xctest"
-               BlueprintName = "litecoin_balanceTests"
-               ReferencedContainer = "container:litecoin_balance.xcodeproj">
+               BuildableName = "LitecoinBalanceTests.xctest"
+               BlueprintName = "LitecoinBalanceTests"
+               ReferencedContainer = "container:LitecoinBalance.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "litecoin_balance.app"
-            BlueprintName = "litecoin_balance"
-            ReferencedContainer = "container:litecoin_balance.xcodeproj">
+            BuildableName = "LitecoinBalance.app"
+            BlueprintName = "LitecoinBalance"
+            ReferencedContainer = "container:LitecoinBalance.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -96,9 +96,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "litecoin_balance.app"
-            BlueprintName = "litecoin_balance"
-            ReferencedContainer = "container:litecoin_balance.xcodeproj">
+            BuildableName = "LitecoinBalance.app"
+            BlueprintName = "LitecoinBalance"
+            ReferencedContainer = "container:LitecoinBalance.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -115,9 +115,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "litecoin_balance.app"
-            BlueprintName = "litecoin_balance"
-            ReferencedContainer = "container:litecoin_balance.xcodeproj">
+            BuildableName = "LitecoinBalance.app"
+            BlueprintName = "LitecoinBalance"
+            ReferencedContainer = "container:LitecoinBalance.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
