+++ ios/LitecoinBalance/AppDelegate.h
