+++ ios/LitecoinBalance/Images-2.xcassets/LaunchImage.launchimage/Contents.json
