+++ ios/LitecoinBalance/Images.xcassets/AppIcon.appiconset/Contents.json
