+++ ios/LitecoinBalance/Images.xcassets/AppIcon.appiconset/Icon-App-20x20@3x.png
