+++ ios/LitecoinBalance/Images.xcassets/AppIcon.appiconset/Icon-App-29x29@3x.png
