+++ ios/LitecoinBalance/Images.xcassets/AppIcon.appiconset/Icon-App-40x40@3x.png
