+++ ios/LitecoinBalance/Images.xcassets/AppIcon.appiconset/Icon-App-60x60@3x.png
