+++ ios/LitecoinBalance/Images.xcassets/AppIcon.appiconset/ItunesArtwork@2x.png
