+++ ios/LitecoinBalance/Images.xcassets/Contents.json
