+++ ios/LitecoinBalance/Images.xcassets/LaunchImage.launchimage/Contents.json
