+++ ios/LitecoinBalance/Images.xcassets/LaunchImage.launchimage/Default1125x2436.png
