+++ ios/LitecoinBalance/Images.xcassets/LaunchImage.launchimage/Default1242x2208.png
