+++ ios/LitecoinBalance/Images.xcassets/LaunchImage.launchimage/Default2208x1242.png
