+++ ios/LitecoinBalance/Images.xcassets/LaunchImage.launchimage/Default2436x1125.png
