+++ ios/LitecoinBalance/Images.xcassets/LaunchImage.launchimage/Default320x480.png
