+++ ios/LitecoinBalance/Images.xcassets/LaunchImage.launchimage/Default640x1136-1.png
