+++ ios/LitecoinBalance/Images.xcassets/LaunchImage.launchimage/Default640x960-1.png
