+++ ios/LitecoinBalance/Images.xcassets/LaunchImage.launchimage/Default640x960.png
