+++ ios/LitecoinBalance/Images.xcassets/LaunchImage.launchimage/Default750x1334.png
