+++ ios/LitecoinBalanceTests/Info.plist
