+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "litecoin_balance",
+  "name": "LitecoinBalance",
   "version": "0.0.1",
   "private": true,
   "scripts": {
